debug_helper.cpp: In member function 'void KBEngine::DebugHelper::sync()': debug_helper.cpp:212: warning: comparison is always false due to limited range of data type

+++ kbe/src/lib/helper/debug_helper.cpp
@@ -205,7 +205,7 @@ void DebugHelper::sync()
 	int8 v = Mercury::g_trace_packet;
 	Mercury::g_trace_packet = 0;
 
-	static int8 trycount = 0;
+	static uint16 trycount = 0;
 	Mercury::Channel* pChannel = pNetworkInterface_->findChannel(messagelogAddr_);
 	if(pChannel == NULL)
 	{
